Follow XDG Base Directory standard
XDG Base Directory is a standard for locations for storing various
files. Nix has a few files which seem to fit in the standard, but
currently use a custom location directly in the user's ~, polluting
it:

- ~/.nix-profile
- ~/.nix-defexpr
- ~/.nix-channels

This commit makes new Nix installations follow the XDG spec and
instead use the following locations:

- $XDG_DATA_HOME/nix/profile
- $XDG_DATA_HOME/nix/defexpr
- $XDG_DATA_HOME/nix/channels

If $XDG_DATA_HOME is not set, it is assumed to be ~/.local/share.

Old locations are preferred if they exist, but for new installations
(or after cleaning the home dir) new, XDG-conformant locations are
used by default.